# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QueueLess Pulse - Smart Queue Management for Banks
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /session/register:
    post:
      operationId: registerUser
      tags: [session]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterInput" }
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /session/login:
    post:
      operationId: loginUser
      tags: [session]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
...
  /branches/{branchId}:
    get:
      operationId: getBranch
      tags: [branches]
      parameters:
        - in: path
          name: branchId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Branch detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BranchPulse" }
...
  /smart-slots:
    get:
      operationId: getSmartSlots
      tags: [bookings]
      parameters:
        - in: query
          name: branchId
          required: true
          schema: { type: integer }
        - in: query
          name: serviceId
          required: true
          schema: { type: integer }
        - in: query
          name: date
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Recommended slots ranked by lowest expected wait
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SmartSlot" }
  /bookings:
    get:
      operationId: listMyBookings
      tags: [bookings]
      responses:
        "200":
          description: My bookings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Booking" }
    post:
      operationId: createBooking
      tags: [bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBookingInput" }
      responses:
        "200":
          description: Booking created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
    delete:
      operationId: cancelBooking
      tags: [bookings]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /bookings/{bookingId}/checklist:
    patch:
      operationId: updateBookingChecklist
      tags: [bookings]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [completedItems]
              properties:
                completedItems:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
...
  /admin/queue/{branchId}:
    get:
      operationId: adminQueueView
      tags: [admin]
      parameters:
        - in: path
          name: branchId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Branch queue view
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminQueueView" }
  /admin/queue/{branchId}/next:
    post:
      operationId: callNextToken
      tags: [admin]
      parameters:
        - in: path
          name: branchId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Next token called
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
  /admin/booking/{bookingId}/status:
    patch:
      operationId: updateBookingStatus
      tags: [admin]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [booked, waiting, serving, completed, no_show, cancelled]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
  /admin/stats/{branchId}:
    get:
      operationId: getBranchStats
      tags: [admin]
      parameters:
        - in: path
          name: branchId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Live throughput stats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BranchStats" }
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    User:
      type: object
      required: [id, name, email, role, karma]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [customer, staff] }
        karma: { type: integer }
    BranchPulse:
      type: object
      required: [id, name, address, city, lat, lng, queueLength, avgWaitMinutes, pulseLevel, openCounters]
      properties:
        id: { type: integer }
        name: { type: string }
        address: { type: string }
        city: { type: string }
        lat: { type: number }
        lng: { type: number }
        queueLength: { type: integer }
        avgWaitMinutes: { type: integer }
        pulseLevel: { type: string, enum: [calm, moderate, busy, packed] }
        openCounters: { type: integer }
        nowServing: { type: string, nullable: true }
...
    SmartSlot:
      type: object
      required: [timeSlot, expectedWaitMinutes, score, label]
      properties:
        timeSlot: { type: string }
        expectedWaitMinutes: { type: integer }
        score: { type: integer }
        label: { type: string }
        recommended: { type: boolean }
    CreateBookingInput:
      type: object
      required: [branchId, serviceId, bookingDate, timeSlot]
      properties:
        branchId: { type: integer }
        serviceId: { type: integer }
        bookingDate: { type: string }
        timeSlot: { type: string }
        groupSize: { type: integer }
    Booking:
      type: object
      required: [id, userId, branchId, branchName, serviceId, serviceName, bookingDate, timeSlot, tokenNumber, status, groupSize, checklistCompleted, checklistTotal, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        branchId: { type: integer }
        branchName: { type: string }
        serviceId: { type: integer }
        serviceName: { type: string }
        bookingDate: { type: string }
        timeSlot: { type: string }
        tokenNumber: { type: string }
        status: { type: string, enum: [booked, waiting, serving, completed, no_show, cancelled] }
        groupSize: { type: integer }
        checklistCompleted: { type: integer }
        checklistTotal: { type: integer }
        createdAt: { type: string }
...
    AdminQueueView:
      type: object
      required: [branch, nowServing, upcoming, completedToday]
      properties:
        branch: { $ref: "#/components/schemas/BranchPulse" }
        nowServing:
          type: array
          items: { $ref: "#/components/schemas/Booking" }
        upcoming:
          type: array
          items: { $ref: "#/components/schemas/Booking" }
        completedToday:
          type: array
          items: { $ref: "#/components/schemas/Booking" }
    BranchStats:
      type: object
      required: [branchId, servedToday, avgServiceMinutes, peakHour, throughputPerHour, hourlyDistribution]
      properties:
        branchId: { type: integer }
        servedToday: { type: integer }
        avgServiceMinutes: { type: integer }
        peakHour: { type: string }
        throughputPerHour: { type: integer }
        hourlyDistribution:
          type: array
          items:
            type: object
            required: [hour, count]
            properties:
              hour: { type: string }
              count: { type: integer }